商品購入機能

+++ app/controllers/items_controller.rb
@@ -20,10 +20,11 @@ class ItemsController < ApplicationController
   end
 
   def show
+    @sale = Sale.all
   end
 
   def edit
-    return unless current_user != @item.user
+    return unless current_user != @item.user || Sale.exists?(item_id: @item.id)
 
     redirect_to root_path
   end
@@ -38,9 +39,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if current_user == item.user
-      item.destroy
-    end
+    item.destroy if current_user == item.user
     redirect_to root_path
   end
